requirements.txt
requests==2.18.4
python-binance==0.6.8
https://github.com/s4w3d0ff/python-poloniex/archive/v0.4.6.zip
